+++ src/path/extra/selectors.ts
@@ -1,0 +1,60 @@
+import { isArray, isObject } from "../../types";
+import { JSONPathEnvironment } from "../environment";
+import { JSONPathNode } from "../node";
+import { JSONPathSelector } from "../selectors";
+import { Token } from "../token";
+
+/**
+ * Object property name selector or array index selector.
+ */
+export class KeysSelector extends JSONPathSelector {
+  constructor(
+    readonly environment: JSONPathEnvironment,
+    readonly token: Token,
+    readonly shorthand: boolean = false,
+  ) {
+    super(environment, token);
+  }
+
+  public resolve(nodes: JSONPathNode[]): JSONPathNode[] {
+    const rv: JSONPathNode[] = [];
+    for (const node of nodes) {
+      if (node.value instanceof String || isArray(node.value)) continue;
+      if (isObject(node.value)) {
+        let i = 0;
+        for (const [key, _] of this.environment.entries(node.value)) {
+          rv.push(
+            new JSONPathNode(
+              key,
+              node.location.concat("[~]", `[${i}]`),
+              node.root,
+            ),
+          );
+          i++;
+        }
+      }
+    }
+    return rv;
+  }
+
+  public *lazyResolve(nodes: Iterable<JSONPathNode>): Generator<JSONPathNode> {
+    for (const node of nodes) {
+      if (node.value instanceof String || isArray(node.value)) continue;
+      if (isObject(node.value)) {
+        let i = 0;
+        for (const [key, _] of this.environment.entries(node.value)) {
+          yield new JSONPathNode(
+            key,
+            node.location.concat("[~]", `[${i}]`),
+            node.root,
+          );
+          i++;
+        }
+      }
+    }
+  }
+
+  public toString(): string {
+    return this.shorthand ? "[~]" : "~";
+  }
+}